mlAttackCheck chance unsigned (cmplw not cmpw)

+++ src/monobj.cpp
@@ -3359,7 +3359,7 @@ int CGMonObj::mlAttackCheck(int partyIndex)
 		if (selectorType == 0) {
 			float maxDist = static_cast<float>(*reinterpret_cast<unsigned short*>(aiScript + actionOffset + 0x114));
 			float minDist = static_cast<float>(static_cast<unsigned int>(*reinterpret_cast<unsigned short*>(aiScript + actionOffset + 0x112)));
-			short chance = *reinterpret_cast<unsigned short*>(aiScript + actionOffset + 0x116);
+			unsigned int chance = *reinterpret_cast<unsigned short*>(aiScript + actionOffset + 0x116);
 
 			if ((targetDist < maxDist) && (minDist < targetDist)) {
